pyproject.toml
[tool.poetry]
name = "sky-link"
version = "0.1.0"
description = ""
authors = ["Rodrigo Almeida <c.almeidarodrigo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-countries = "^7.6.1"
factory-boy = "^3.3.1"
faker = "^30.8.2"
tqdm = "^4.67.0"
drf-spectacular = "^0.27.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
